Renamed files to lower-case starting letter
