+++ src/components/common/cache/cache.test.ts
@@ -1,4 +1,4 @@
-import { Cache } from './Cache'
+import { Cache } from './cache'
 
 describe('Test caching functionality', () => {
   let testCache: Cache<string, number>
